+++ src/resources/PortabilityChecks.ts
@@ -3,6 +3,7 @@ const telnyxMethod = TelnyxResource.method;
 
 export const PortabilityChecks = TelnyxResource.extend({
   path: 'portability_checks',
+  includeBasic: ['create'],
 
   run: telnyxMethod({
     method: 'POST',