+++ src/resources/PortingReports.ts
@@ -1,0 +1,6 @@
+import TelnyxResource from '../TelnyxResource';
+
+export const PortingReports = TelnyxResource.extend({
+  path: 'porting/reports',
+  includeBasic: ['list', 'create', 'retrieve'],
+});